pyproject.toml
[tool.poetry]
name = "weather-data-downloader"
version = "0.1.0"
description = "A scalable weather data downloader for numerical weather prediction models"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
wd = "src.cli.main:cli"

[tool.poetry.dependencies]
python = "^3.12"
requests = ">=2.32.5,<3.0.0"
pandas = ">=2.3.2,<3.0.0"
numpy = ">=2.3.2,<3.0.0"
click = ">=8.2.1,<9.0.0"
rich = ">=14.1.0,<15.0.0"
pydantic = ">=2.11.7,<3.0.0"
PyYAML = ">=6.0.1,<7.0.0"
loguru = "^0.7.3"
xarray = "^2025.8.0"
cfgrib = "^0.9.15.0"
netcdf4 = "^1.7.2"
dask = "^2025.7.0"
scipy = "^1.16.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
netcdf4 = "^1.7.2"
xarray = "^2025.8.0"
matplotlib = "^3.10.5"
cartopy = "^0.25.0"
cfgrib = "^0.9.15.0"
aiohttp = "^3.12.15"
asyncio-throttle = "^1.0.2"
tenacity = "^9.1.2"
commitizen = "^4.8.3"
pre-commit = "^4.3.0"
flake8 = "^7.3.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.0"
responses = "^0.25.0"
freezegun = "^1.5.0"
factory-boy = "^3.3.0"
hypothesis = "^6.98.0"
pytest-xdist = "^3.5.0"
pytest-html = "^4.1.0"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "--strict-markers",
    "--strict-config", 
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=85",
    "-ra",
    "--tb=short",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Slow tests",
    "network: Tests requiring network",
    "cli: CLI command tests",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version"
]
bump_message = "bump: version $current_version → $new_version"
changelog_file = "CHANGELOG.md"
changelog_format = "## $version ($date)\n\n$changes\n"
change_type_order = ["BREAKING CHANGE", "feat", "fix", "refactor", "perf"]

